[REFACTOR-019] decompose LosTab — extract los-admin-types, VitalSignsConfig; lift expand state; apiClient migration
